refactor(clawdstrike): extract posture impl from engine.rs
Move the posture-aware entrypoints (check_action_report_with_posture,
check_action_report_with_runtime) and the runtime posture state machine
(ensure_posture_initialized, normalize_state_budgets,
apply_timeout_transitions, posture_precheck, apply_trigger_transition,
apply_transition) plus the PosturePrecheck helper type into posture.rs as
an impl HushEngine block.

The block calls the already-pub(crate) decision-flow methods; no further
widening needed. Test-only types (Error, PostureRuntimeState,
PostureTransitionRecord) moved to tests/mod.rs shared imports. Public API
unchanged.

Pure relocation: 825 lib tests + formal-diff-tests pass.
